requirements.txt
beautifulsoup4==4.12.2
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-tools==1.53.0
more-itertools==9.1.0
protobuf==4.22.1
PyYAML==6.0
requests==2.28.2